move some headers to internal
