+++ sdk/storage/azure-storage-blobs/src/append_blob_client.cpp
@@ -3,7 +3,7 @@
 
 #include "azure/storage/blobs/append_blob_client.hpp"
 
-#include <azure/storage/common/constants.hpp>
+#include <azure/storage/common/internal/constants.hpp>
 #include <azure/storage/common/storage_common.hpp>
 
 namespace Azure { namespace Storage { namespace Blobs {
