+++ sdk/storage/azure-storage-blobs/test/ut/block_blob_client_test.cpp
@@ -9,7 +9,7 @@
 
 #include <azure/core/cryptography/hash.hpp>
 #include <azure/storage/common/crypt.hpp>
-#include <azure/storage/common/file_io.hpp>
+#include <azure/storage/common/internal/file_io.hpp>
 
 namespace Azure { namespace Storage { namespace Blobs { namespace Models {
