+++ sdk/storage/azure-storage-common/inc/azure/storage/common/internal/file_io.hpp
