+++ sdk/storage/azure-storage-common/inc/azure/storage/common/internal/reliable_stream.hpp
