+++ sdk/storage/azure-storage-common/inc/azure/storage/common/internal/shared_key_policy.hpp
