+++ sdk/storage/azure-storage-common/inc/azure/storage/common/internal/xml_wrapper.hpp
