+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_client.cpp
@@ -4,9 +4,9 @@
 #include "azure/storage/files/datalake/datalake_file_client.hpp"
 
 #include <azure/core/http/policies/policy.hpp>
-#include <azure/storage/common/constants.hpp>
 #include <azure/storage/common/crypt.hpp>
-#include <azure/storage/common/shared_key_policy.hpp>
+#include <azure/storage/common/internal/constants.hpp>
+#include <azure/storage/common/internal/shared_key_policy.hpp>
 #include <azure/storage/common/storage_common.hpp>
 
 #include "azure/storage/files/datalake/datalake_constants.hpp"
