+++ sdk/storage/azure-storage-files-shares/test/share_file_client_test.cpp
@@ -9,7 +9,7 @@
 
 #include <azure/core/cryptography/hash.hpp>
 #include <azure/storage/common/crypt.hpp>
-#include <azure/storage/common/file_io.hpp>
+#include <azure/storage/common/internal/file_io.hpp>
 #include <azure/storage/common/storage_common.hpp>
 
 namespace Azure { namespace Storage { namespace Test {
